+++ CHANGELOG.md
@@ -3,6 +3,8 @@
 ## Version [7.0.0](https://github.com/cedx/coveralls.php/compare/v6.0.0...v7.0.0)
 - Breaking change: raised the required [PHP](https://secure.php.net) version.
 - Breaking change: using PHP 7.1 features, like nullable types and void functions.
+- Added the `ClientException` class.
+- Added a user guide based on [MkDocs](http://www.mkdocs.org).
 - Updated the package dependencies.
 
 ## Version [6.0.0](https://github.com/cedx/coveralls.php/compare/v5.0.0...v6.0.0)